[Refacotr] 프로토콜 별도의 파일로 분리
- 가독성 향상을 위해 파일 분리
